Remove stub pacakge generation
The stub package generation was from a previous model of the build
where we needed to ensure host packages for all supported platforms
existed when we restored test assets.

We no longer have this dependency, we only ever consume packages from
the current platform we are building, so we can remove all of this
code.